requirements.txt
mysql-connector-python==8.1.0
TTS==0.21.3
